requirements.txt
EasyProcess==0.2.3
PyVirtualDisplay==0.2.1
PyYAML==3.12
selenium==3.8.0
mysqlclient==1.3.12
requests==2.18.4
urllib3==1.22
